requirements-dev.txt
# Development and Testing Dependencies
# =====================================

# Testing Framework
pytest>=8.2
pytest-asyncio>=0.23
pytest-cov>=5.0

# Aiogram Testing (Official)
aiogram-tests==1.0.3

# Mocking and Fixtures
fakeredis[lua]==2.*
respx>=0.21
httpx>=0.27

# Code Quality (Optional - for future)
# pylint>=3.0
# mypy>=1.8
# black>=24.0
# isort>=5.13
